WIP - everything is hooked up but just learned that priceInsights needs internalID instead of slug
