+++ src/lib/Components/Artist/ArtistInsights/ArtistInsights.tsx
@@ -1,20 +1,22 @@
-import { ActionType, ContextModule, OwnerType, tappedInfoBubble, TappedInfoBubbleArgs } from "@artsy/cohesion"
+import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
 import { ArtistInsights_artist } from "__generated__/ArtistInsights_artist.graphql"
-import { InfoButton } from "lib/Components/Buttons/InfoButton"
+import { ArtistInsights_priceInsights } from "__generated__/ArtistInsights_priceInsights.graphql"
 import { AnimatedArtworkFilterButton, FilterModalMode, FilterModalNavigator } from "lib/Components/FilterModal"
 import { StickyTabPageScrollView } from "lib/Components/StickyTabPage/StickyTabPageScrollView"
 import { ArtworkFilterGlobalStateProvider } from "lib/utils/ArtworkFilter/ArtworkFiltersStore"
 import { useScreenDimensions } from "lib/utils/useScreenDimensions"
-import { Flex, Join, Separator, Spacer, Text } from "palette"
+import { Join, Separator } from "palette"
 import React, { useCallback, useState } from "react"
-import { NativeScrollEvent, NativeSyntheticEvent, ScrollView } from "react-native"
+import { NativeScrollEvent, NativeSyntheticEvent } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
 import { useTracking } from "react-tracking"
 import { ReactElement } from "simple-markdown"
 import { ArtistInsightsAuctionResultsPaginationContainer } from "./ArtistInsightsAuctionResults"
+import { MarketStatsFragmentContainer } from "./MarketStats"
 
 interface ArtistInsightsProps {
   artist: ArtistInsights_artist
+  priceInsights: ArtistInsights_priceInsights
 }
 
 export interface ViewableItems {
@@ -30,9 +32,8 @@ interface ViewToken {
 }
 
 const FILTER_BUTTON_OFFSET = 50
-export const ArtistInsights: React.FC<ArtistInsightsProps> = ({ artist }) => {
+export const ArtistInsights: React.FC<ArtistInsightsProps> = ({ artist, priceInsights }) => {
   const tracking = useTracking()
-
   const [isFilterButtonVisible, setIsFilterButtonVisible] = useState(false)
   const [isFilterModalVisible, setIsFilterModalVisible] = useState(false)
 
@@ -55,88 +56,6 @@ export const ArtistInsights: React.FC<ArtistInsightsProps> = ({ artist }) => {
     setIsFilterButtonVisible(false)
   }, [])
 
-  const renderInfoModal = () => (
-    <ScrollView>
-      <Text>
-        The following data points provide an overview of an artist’s auction market for a specific medium (e.g.,
-        photography, painting) over the past 36 months.
-      </Text>
-      <Spacer mb={2} />
-      <Text>
-        These market signals bring together data from top auction houses around the world, including Christie’s,
-        Sotheby’s, Phillips, Bonhams, and Heritage.
-      </Text>
-      <Spacer mb={2} />
-      <Text>
-        In this data set, please note that the sale price includes the hammer price and buyer’s premium, as well as any
-        other additional fees (e.g., Artist’s Resale Rights).
-      </Text>
-      <Spacer mb={2} />
-      <Text fontWeight={"bold"}>Yearly lots sold</Text>
-      <Spacer mb={1} />
-      <Text>The average number of lots sold per year at auction over the past 36 months.</Text>
-      <Spacer mb={2} />
-      <Text fontWeight={"bold"}>Sell-through rate</Text>
-      <Spacer mb={1} />
-      <Text>The percentage of lots in auctions that sold over the past 36 months.</Text>
-      <Spacer mb={2} />
-      <Text fontWeight={"bold"}>Average sale price</Text>
-      <Spacer mb={1} />
-      <Text>The average sale price of lots sold at auction over the past 36 months.</Text>
-      <Spacer mb={2} />
-      <Text fontWeight={"bold"}>Sale price over estimate</Text>
-      <Spacer mb={1} />
-      <Text lineHeight={50}>
-        The average percentage difference of the sale price over the mid-estimate (the midpoint of the low and high
-        estimates set by the auction house before the auction takes place) for lots sold at auction over the past 36
-        months.
-      </Text>
-      <Spacer mb={2} />
-    </ScrollView>
-  )
-
-  const renderMarketStats = () => (
-    <>
-      {/* Market Stats Hint */}
-      <Flex flexDirection="row" alignItems="center">
-        <InfoButton
-          titleElement={
-            <Text variant="title" mr={0.5}>
-              Market Signals by Medium
-            </Text>
-          }
-          trackEvent={() => {
-            tracking.trackEvent(tappedInfoBubble(tracks.tapMarketStatsInfo()))
-          }}
-          modalTitle={"Market Signals by Medium"}
-          modalContent={renderInfoModal()}
-        />
-      </Flex>
-      <Text variant="small" color="black60">
-        Last 12 months
-      </Text>
-      {/* Market Stats Values */}
-      <Flex flexDirection="row" flexWrap="wrap" mt={15}>
-        <Flex width="50%">
-          <Text variant="text">Average sale price</Text>
-          <Text variant="largeTitle">$168k</Text>
-        </Flex>
-        <Flex width="50%">
-          <Text variant="text">Total lots sold</Text>
-          <Text variant="largeTitle">61</Text>
-        </Flex>
-        <Flex width="50%" mt={2}>
-          <Text variant="text">Realized / estimate</Text>
-          <Text variant="largeTitle">2.12x</Text>
-        </Flex>
-        <Flex width="50%" mt={2}>
-          <Text variant="text">Sell-through rate</Text>
-          <Text variant="largeTitle">90%</Text>
-        </Flex>
-      </Flex>
-    </>
-  )
-
   return (
     <ArtworkFilterGlobalStateProvider>
       <StickyTabPageScrollView
@@ -144,7 +63,7 @@ export const ArtistInsights: React.FC<ArtistInsightsProps> = ({ artist }) => {
         onScrollEndDrag={onScrollEndDrag}
       >
         <Join separator={<Separator my={2} ml={-2} width={useScreenDimensions().width} />}>
-          {renderMarketStats()}
+          <MarketStatsFragmentContainer priceInsights={priceInsights} />
           <ArtistInsightsAuctionResultsPaginationContainer artist={artist} />
         </Join>
       </StickyTabPageScrollView>
@@ -176,6 +95,11 @@ export const ArtistInsightsFragmentContainer = createFragmentContainer(ArtistIns
       ...ArtistInsightsAuctionResults_artist
     }
   `,
+  priceInsights: graphql`
+    fragment ArtistInsights_priceInsights on PriceInsightConnection {
+      ...MarketStats_priceInsights
+    }
+  `,
 })
 
 export const tracks = {
@@ -199,10 +123,4 @@ export const tracks = {
       action_type: ActionType.commercialFilterParamsChanged,
     }
   },
-
-  tapMarketStatsInfo: (): TappedInfoBubbleArgs => ({
-    contextModule: ContextModule.auctionResults,
-    contextScreenOwnerType: OwnerType.artistInsights,
-    subject: "artistMarketStatistics",
-  }),
 }
