+++ src/lib/Scenes/Artist/Artist.tsx
@@ -28,7 +28,8 @@ import { RelayModernEnvironment } from "relay-runtime/lib/store/RelayModernEnvir
 export const Artist: React.FC<{
   artistAboveTheFold: NonNullable<ArtistAboveTheFoldQuery["response"]["artist"]>
   artistBelowTheFold?: ArtistBelowTheFoldQuery["response"]["artist"]
-}> = ({ artistAboveTheFold, artistBelowTheFold }) => {
+  priceInsights?: ArtistBelowTheFoldQuery["response"]["priceInsights"]
+}> = ({ artistAboveTheFold, artistBelowTheFold, priceInsights }) => {
   const tabs = []
   const displayAboutSection =
     artistAboveTheFold.has_metadata ||
@@ -62,7 +63,12 @@ export const Artist: React.FC<{
   if (isArtistInsightsEnabled && artistAboveTheFold?.auctionResultsConnection?.totalCount) {
     tabs.push({
       title: "Insights",
-      content: artistBelowTheFold ? <ArtistInsightsFragmentContainer artist={artistBelowTheFold} /> : <LoadingPage />,
+      content:
+        artistBelowTheFold && priceInsights ? (
+          <ArtistInsightsFragmentContainer artist={artistBelowTheFold} priceInsights={priceInsights} />
+        ) : (
+          <LoadingPage />
+        ),
     })
   }
 
@@ -104,6 +110,7 @@ interface ArtistQueryRendererProps extends ArtistAboveTheFoldQueryVariables, Art
 }
 
 export const ArtistQueryRenderer: React.FC<ArtistQueryRendererProps> = ({ artistID, environment }) => {
+  console.log("sjhsjhsjh", artistID)
   return (
     <AboveTheFoldQueryRenderer<ArtistAboveTheFoldQuery, ArtistBelowTheFoldQuery>
       environment={environment || defaultEnvironment}
@@ -130,17 +137,21 @@ export const ArtistQueryRenderer: React.FC<ArtistQueryRendererProps> = ({ artist
         `,
         variables: { artistID },
       }}
+      // uh oh. artist ID is the "id" field; priceInsights needs the "internalID" field. How & where can we convert this?
       below={{
         query: graphql`
-          query ArtistBelowTheFoldQuery($artistID: String!, $isPad: Boolean!) {
-            artist(id: $artistID) {
+          query ArtistBelowTheFoldQuery($artistID: ID!, $artistIDString: String!, $isPad: Boolean!) {
+            artist(id: $artistIDString) {
               ...ArtistAbout_artist
               ...ArtistShows_artist
               ...ArtistInsights_artist
             }
+            priceInsights(artistId: $artistID) {
+              ...ArtistInsights_priceInsights
+            }
           }
         `,
-        variables: { artistID, isPad: isPad() },
+        variables: { artistID, artistIDString: artistID, isPad: isPad() },
       }}
       render={{
         renderPlaceholder: () => <HeaderTabsGridPlaceholder />,
@@ -148,7 +159,13 @@ export const ArtistQueryRenderer: React.FC<ArtistQueryRendererProps> = ({ artist
           if (!above.artist) {
             throw new Error("no artist data")
           }
-          return <Artist artistAboveTheFold={above.artist} artistBelowTheFold={below?.artist} />
+          return (
+            <Artist
+              artistAboveTheFold={above.artist}
+              artistBelowTheFold={below?.artist}
+              priceInsights={below?.priceInsights}
+            />
+          )
         },
       }}
     />
